impl verify
