+++ mopro-ffi/Cargo.toml
@@ -45,7 +45,6 @@ color-eyre = "=0.6.2"
 ashlang = { git = "https://github.com/chancehudson/ashlang.git", rev = "696960a0c15db47170fdd1ff058682023d904b1b", optional = true, default-features = false, features = ["spartan-prover", "serde"] }
 
 # circom deps
-rust-witness = { version = "0.1.1", optional = true }
 ark-circom = { git = "https://github.com/zkmopro/circom-compat.git", version = "0.1.0", branch = "wasm-delete", optional = true }
 ark-serialize = { version = "=0.4.1", features = ["derive"], optional = true }
 num-bigint = { version = "0.4.3", default-features = false, features = [
